requirements.txt
Django==6.0.3
djangorestframework==3.14.0
dj-rest-auth==5.0.2
dj-rest-auth[with_social]==5.0.2
django-allauth==0.57.0
Pillow==10.1.0
python-decouple==3.8
requests==2.31.0
google-genai==0.5.2
python-dotenv==1.0.0
